Provide completions on a per machine basis
make (Makefile) gets build information from two (2) files:

    GENERATE
        targets to be built

    SHARED
        targets to be soft linked from $KDOTDIR/share/CS

In addition to editing these files to include completions you want, in
the case of GENERATE, you will also have to edit Makefile and add the
target and build instructions.

The Makefile tests the directory it's being run in and will exit without
doing anything if the machine it is being run on $(HOST) does not match
the directory it is being run in ($KDOTDIR/$HOST/*). All edits MUST
conform to this.

+++ share/BS/update-help-completions.ksh
@@ -4,9 +4,9 @@
 
 set -o nounset;: ${FPATH:?Run from within KSH} ${KDOTDIR:?}
 
-BASED=$KDOTDIR/share
-CmplHelp=$BASED/CS/help
-HelpDir=$BASED/HS
+BASED=$KDOTDIR/$HOST
+CmplHelp=$BASED/C/help
+HelpDir=$BASED/H
 
 # Usage {{{1
 typeset -- this_pgm="${0##*/}"
@@ -48,21 +48,24 @@ function warnOrDie { #{{{1
 	esac
 } # }}}1
 
-
 TEMPF=$CmplHelp.tmp
 
 o=$IFS; set -- $FPATH; IFS=$o
-for p in $HelpDir "$@"; do printf "%s\n" "$p"/*; done \
-	| sed -e 's+.*/++' -e '/\*$/d'					\
-	| sort											\
-	| uniq											\
-	> $TEMPF
+for p in $HelpDir "$@"; do printf "%s\n" "$p"/*; done	\
+	| sed -e 's+.*/++' -e '/\*$/d'						\
+	| sort												\
+	| uniq												\
+	>$TEMPF
 
-cmp -s $CmplHelp $TEMPF || { 
-	cat $TEMPF >$CmplHelp
-	print help # for Makefile output
-  }
-
-rm $TEMPF
+if [[ -f $CmplHelp ]]; then
+	cmp -s $CmplHelp $TEMPF || { 
+		cat $TEMPF >$CmplHelp
+		print help # for Makefile output
+	  }
+	rm $TEMPF
+else # $CmplHelp doesn't exist
+	mv $TEMPF $CmplHelp
+	print +help # for Makefile output
+fi
 
 # Copyright (C) 2017 by Tom Davis <tom@greyshirt.net>.
